MDCACHE - Fix chunk refcounting in readdir
Chunk refcounting was a bit loose, and there was a case where a chunk
could get reaped while in use, when the write lock was held.

Clean up the chunk refcounting, so that all the pointers to chunks that
we keep during readdir have a refcount, and that we drop all the refs
appropriately.

Resolves: rhbz#1708587
Resolves: rhbz#1715954

Change-Id: Ic3acfc71d15cde1cdf68a46864f5e60445c4ff71
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>