MDCACHE - add a name-cookie mapping for whence-is-name
Restarting a readdir internally because the chunk containing the
requested cookie has been reaped cause extreme performance problems on
large directories.  It can cause a hang long enough for the client to
time out.

To avoid this, add a name-cookie map.  This map contains the last cookie
we sent to a client, so that when that cookie is sent back by the
client, we can look up the name even if the chunk containing that dirent
has been reaped.  This map is in an LRU, with a hard cap
(Dirmap_HWMark), and it's cleaned up in the background if entries are
older than 60 seconds.

Notet that we also save the second cookie in the set, since the linux
client seems to request that one approximately 1/3 of the time.

Resolves: rhbz#1708587
Resolves: rhbz#1715954

Change-Id: I09d1ef6012d5feca81c7418634cba02e3ab08bc5
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>