FSAL: Fix up how reaper_work is used
While mdcache uses multiple lanes of LRU to track entries and handle
reaping of them, and it used to do the global fd management, now that
FSAL/commonlib.c does the fd_lru, it is a single lru queue, so it
will not do per lane calculations.

So the old calculation that calculated a per lane work, is now just
reaper_work and that has to be properly set and used.

Change-Id: Ife6ce9822edfdecb34d2cfb4b5a264bae95ad4ff
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>