CONFIG: %plugin architectyure
This is a mechanism to load in code that will (eventually)
be capable of reading custom configuration and using those
values to augment the actions of fsals on a per-export basis.

My initial plans here are specifically to use kmip to fetch a
key to use with fscrypt and cephfs, but this mechanism is
intended to be more generic.

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit bd5d7fd0032b6bf08b558cbf6f9e71dc5cc2a969)