fsal control: Add "control" operation for fsal object.
Two operations are defined: FSCRYPT_SETKEY and
FSCRYPT_VERIFY_NOT_ENCRYPTED, both are for fscrypt support.

This is just the plumbing, it is up to individual
FSALs to implement the actual logic.  However, an important
piece of the plumbing that is included here is mdcache logic
to pass operations to the underlying FSAL.

Change-Id: I065dd583e8ec4b6c155bc816598879f83c64a1f8
Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 3739cffb0973b183a5c5404cdb0e00f1308a6b4b)