plugin modules - use module directory not cwd.
Turns out dlopen doesn't do anything useful with
"ganesha/libFOO.so".  Also to harmonize with other
prats of ganesh that should be just "FOO" not "ganesha/libFOO.so".

So this adds logic to build ig against the module location and also
work with "ganesha/libFOO.so" since I had promised this would work.

Currently this is called before the global block is parsed;
it hsould be called later so that the module location can
be customized.

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 1597451202a456ccb9a14ca06f665edcd976dbd6)