fscrypt: initial export should not ignore extension errors.
If an "export root callback" fails, that indicates that there was a
problem with an extension fo an export, and the export should not be
made available.  This is already true when updating an export, this
commit makes it so for the initial export creation too.

XXX merge with previous

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit fbcc1e6b544e67a1a1a59c9f8ae474a9448711d8)